+++ test/conftest.py
@@ -62,24 +62,24 @@ def pytest_report_header(config):
 
 def pytest_generate_tests(metafunc):
 
-    if 'fast_math' in metafunc.funcargnames:
+    if 'fast_math' in metafunc.fixturenames:
         fm = metafunc.config.option.fast_math
         fms = dict(both=[False, True], no=[False], yes=[True])[fm]
         fm_ids = [{False:'nofm', True:'fm'}[fm] for fm in fms]
         metafunc.parametrize('fast_math', fms, ids=fm_ids)
 
-    if 'thr_and_double' in metafunc.funcargnames:
+    if 'thr_and_double' in metafunc.fixturenames:
         parametrize_thread_tuple(metafunc, 'thr_and_double', pair_thread_with_doubles)
 
-    if 'thr' in metafunc.funcargnames:
+    if 'thr' in metafunc.fixturenames:
         tps, tp_ids = get_threads(metafunc.config)
         metafunc.parametrize('thr', tps, ids=tp_ids, indirect=True)
 
-    if 'some_thr' in metafunc.funcargnames:
+    if 'some_thr' in metafunc.fixturenames:
         # Just some thread for tests that only check thread-independent stuff.
         tps, tp_ids = get_threads(metafunc.config)
         metafunc.parametrize('some_thr', [tps[0]], ids=[tp_ids[0]], indirect=True)
 
-    if 'cluda_api' in metafunc.funcargnames:
+    if 'cluda_api' in metafunc.fixturenames:
         apis, api_ids = get_apis(metafunc.config)
         metafunc.parametrize('cluda_api', apis, ids=api_ids)