+++ test/test_cbrng/test_cbrng.py
@@ -115,7 +115,7 @@ class BijectionHelper:
     def reference(self, counters, keygen):
         result = numpy.empty_like(counters)
         for i in range(counters.shape[0]):
-            result[i] = self._reference_func(counters[i]['v'], keygen(i)['v'])
+            result[i]['v'] = self._reference_func(counters[i]['v'], keygen(i)['v'])
         return result
 
     def __str__(self):
@@ -125,7 +125,7 @@ class BijectionHelper:
 
 def pytest_generate_tests(metafunc):
 
-    if 'test_bijection' in metafunc.funcargnames:
+    if 'test_bijection' in metafunc.fixturenames:
 
         vals = []
         ids = []
@@ -137,13 +137,13 @@ def pytest_generate_tests(metafunc):
 
         metafunc.parametrize('test_bijection', vals, ids=ids)
 
-    if 'test_sampler_int' in metafunc.funcargnames:
+    if 'test_sampler_int' in metafunc.fixturenames:
         vals = [
             UniformIntegerHelper(-10, 98)]
         ids = [test.name for test in vals]
         metafunc.parametrize('test_sampler_int', vals, ids=ids)
 
-    if 'test_sampler_float' in metafunc.funcargnames:
+    if 'test_sampler_float' in metafunc.fixturenames:
         vals = [
             UniformFloatHelper(-5, 7.7),
             NormalBMHelper(-2, 10),